% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNearest.R
\name{assembly}
\alias{assembly}
\title{Get the required information for the given assembly}
\usage{
assembly(
  org_assembly = c("hg19", "hg38", "mm10", "dre10", "rn6", "dm6", "ce11", "sc3")
)
}
\arguments{
\item{org_assembly}{Genome assembly of interest for the analysis. Possible
assemblies are "mm10" for mouse, "dre10" for zebrafish, "rn6" for rat,
"dm6" for fruit fly, "ce11" for worm, "sc3" for yeast, "hg19" and "hg38"
for human}
}
\value{
setting required information
}
\description{
Get the required information for the given assembly
}
\examples{
 \dontrun{
assembly('hg19') 
}

}
