% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{brain_disorder_ncRNA}
\alias{brain_disorder_ncRNA}
\title{Differentially expressed non-coding gene}
\format{
Not Available
}
\source{
\url{http://resource.psychencode.org/}
}
\usage{
brain_disorder_ncRNA
}
\description{
Differentially expressed non-coding gene
}
\examples{
data(brain_disorder_ncRNA)
}
\keyword{datasets}
