% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{brain_mirna}
\alias{brain_mirna}
\title{Differentially expressed human brain data}
\format{
Not Available
}
\source{
\url{http://resource.psychencode.org/}
}
\usage{
brain_mirna
}
\description{
Differentially expressed human brain data
}
\examples{
data(brain_mirna)
}
\keyword{datasets}
