% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{breastmRNA}
\alias{breastmRNA}
\title{Protein coding genes that are differentially expressed in TCGA breast cancer RNAseq data.}
\format{Not Available}
\source{
\url{https://portal.gdc.cancer.gov/}
}
\usage{
breastmRNA
}
\description{
Protein coding genes that are differentially expressed in TCGA breast cancer RNAseq data.
}
\examples{
data(breastmRNA)
}
\keyword{datasets}
