% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{convertGMT}
\alias{convertGMT}
\title{Convert gmt formatted pathway file to the Pathway ID, Entrez, symbol
 formatted data frame}
\usage{
convertGMT(gmtName, isSymbol = FALSE)
}
\arguments{
\item{gmtName}{Custom pathway gmt file}

\item{isSymbol}{Boolean variable that hold the gene format of the gmt file.
If it is set as TRUE, gene format of the gmt file should be symbol.
Otherwise, gene format should be ENTREZ ID. By default, it is FALSE.}
}
\value{
return data frame
}
\description{
Convert gmt formatted pathway file to the Pathway ID, Entrez, symbol
 formatted data frame
}
