% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionBased.R
\name{corrbasedMrna}
\alias{corrbasedMrna}
\title{Pearson correlation coefficient value of the mRNA genes between miRNA:mRNA
for a given correlation cut-off and cancer.}
\usage{
corrbasedMrna(mRNAgene, cancer, minAbsCor, databaseFile)
}
\arguments{
\item{mRNAgene}{Data frame of the mRNA genes}

\item{cancer}{Name of the TCGA project code such as 'BRCA' that is analyzed
for miRNA-mRNA correlation. Possible cancer types ACC, BLCA, BRCA, 
CESC, CHOL, COAD, COADREAD, DLBC, ESCA, GBMLGG, HNSC, KICH, KIPAN, 
KIRC, KIRP, LGG, LIHC, LUAD, LUSC, OV, PAAD, PCPG, PRAD, READ, SARC, 
SKCM, STAD, STES, TGCT, THCA, THYM, UCEC, UCS, UVM}

\item{minAbsCor}{Cut-off value for the Pearson correlation coefficient of
the miRNA-mRNA}

\item{databaseFile}{Path of miRcancer.db file}
}
\value{
Data frame of the miRNA-mRNA correlation result
}
\description{
Pearson correlation coefficient value of the mRNA genes between miRNA:mRNA
for a given correlation cut-off and cancer.
}
\examples{

database<- "filePath//miRCancer.db"
\dontrun{
d<-corrbasedMrna(mRNAgene = as.data.frame(gene_list$geneSet),
                 cancer = 'BRCA', minAbsCor = 0.3, database)
}

}
