% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{drawDotPlot}
\alias{drawDotPlot}
\title{Draw dot plot of the enrichment object}
\usage{
drawDotPlot(mrnaObject, type = "pAdjust", n)
}
\arguments{
\item{mrnaObject}{Object of the enrichment result}

\item{type}{Draw the dot plot according to the p-value or adjusted p-value
("pvalue", "pAdjust")}

\item{n}{Number of GO terms or pathways, that ordered by type and has
least number of top p-value}
}
\value{
Dot plot of the top n enrichment results
}
\description{
Draw dot plot of the enrichment object
}
\examples{
ncGO <- geneGOEnricher(gene = brain_disorder_ncRNA[1:100,],
  org_assembly='hg19', near=TRUE,  genetype = 'Ensembl_gene')

}
