% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{fly_c}
\alias{fly_c}
\title{GO-term annotation database from the cellular component ontology of the fly genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
fly_c
}
\description{
GO-term annotation database from the cellular component ontology of the fly genes
}
\examples{
data(fly_c)
}
\keyword{datasets}
