% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{fly_f}
\alias{fly_f}
\title{GO-term annotation database from the molecular function ontology of the fly genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
fly_f
}
\description{
GO-term annotation database from the molecular function ontology of the fly genes
}
\examples{
data(fly_f)
}
\keyword{datasets}
