% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{fly_p}
\alias{fly_p}
\title{GO-term annotation database from the biological process ontology of the fly genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
fly_p
}
\description{
GO-term annotation database from the biological process ontology of the fly genes
}
\examples{
data(fly_p)
}
\keyword{datasets}
