% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{gencode}
\alias{gencode}
\title{GENCODE v30 whole gene set}
\format{Not Available}
\source{
\url{https://www.gencodegenes.org/}
}
\usage{
gencode
}
\description{
GENCODE v30 whole gene set
}
\examples{
data(gencode)
}
\keyword{datasets}
