% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEnrichment.R
\name{geneRegionPathwayEnricher}
\alias{geneRegionPathwayEnricher}
\title{Given gene regions that fall in the given upstream and downstream region of
mRNAs of interest, pathway enrichment analysis is carried out}
\usage{
geneRegionPathwayEnricher(region, org_assembly = c("hg19", "hg38",
  "mm10", "dre10", "rn6", "dm6", "ce11", "sc3"), upstream = 10000,
  downstream = 10000, searchRegion = c("all", "exon", "intron"),
  pCut = 0.05, pAdjCut = 0.05, pAdjust = c("holm", "hochberg",
  "hommel", "bonferroni", "BH", "BY", "fdr", "none"), min = 5,
  pathwayType = c("kegg", "reactome", "wiki", "other"), near = FALSE,
  isTADSearch = FALSE, TAD = tad_hg19, cellline = "all",
  gmtName = "", isSymbol = FALSE, express = FALSE,
  isCustomExp = FALSE, cancer, exp1, exp2, label1 = "", label2 = "",
  corrMethod = c("pearson", "kendall", "spearman"), varCutoff = 0.0025,
  minAbsCor = 0.3, pcut = 0.05, alternate = c("greater", "two.sided",
  "less"), isUnionCorGene = FALSE, conf = 0.95, databaseFile,
  isGeneEnrich = FALSE)
}
\arguments{
\item{region}{Bed format of input gene regions other than miRNA. Input must
be Granges object.}

\item{org_assembly}{Genome assembly of interest for the analysis. Possible
assemblies are "mm10" for mouse, "dre10" for zebrafish, "rn6" for rat, 
"dm6" for fruit fly, "ce11" for worm, "sc3" for yeast, "hg19" and 
"hg38" for human}

\item{upstream}{Upstream distance from the transcription start position}

\item{downstream}{Downstream distance from the transcription end position}

\item{searchRegion}{Search space of the cis-region. Possible values are 
"all", "exon", "intron"}

\item{pCut}{Threshold value for the pvalue. Default value is 0.05}

\item{pAdjCut}{Cutoff value for the adjusted p-values using one of given
method. Default value is 0.05.}

\item{pAdjust}{Methods of the adjusted p-values. Possible methods are
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr", "none"}

\item{min}{Minimum number of genes that are required for enrichment. By 
default, it is set to 5.}

\item{pathwayType}{Pathway database for enrichment. Possible values are 
'reactome' for Reactome, 'kegg' for KEGG, 'wiki' for WikiPathways, 
'other' for custom database}

\item{near}{Boolean value presents whether cis-neighbourhood should be
considered in the analysis}

\item{isTADSearch}{Boolean value that shows whether TAD analysis is 
performed. This value has to be TRUE for TAD analysis.}

\item{TAD}{TAD genomic regions for the species. Predefined TAD regions or
any new TAD regions can be used for the analysis. TAD regions must be 
formated as GRanges object. Predefined TAD regions are 'tad_hg19', 
'tad_hg38', 'tad_mm10', 'tad_dmel' for hg19, hg38, mm9 and dm6 assembly,
respectively.}

\item{cellline}{Cell lines for TAD regions.}

\item{gmtName}{Custom pathway gmt file}

\item{isSymbol}{Boolean variable that hold the gene format of the gmt file. 
If it is set as TRUE, gene format of the gmt file should be symbol. 
Otherwise, gene format should be ENTREZ ID. By default, it is FALSE.}

\item{express}{Boolean variable whether co-expression analysis is 
performed. If this option is set to TRUE, co-expression analysis will
be performed.}

\item{isCustomExp}{Boolean variable whether co-expression analysis with 
custom data will be performed. When this option is set, exp1 and exp2
parameters must be defined.}

\item{cancer}{Defines the name of the TCGA project code such as 'BRCA' for
correlation analysis. Possible cancer types ACC, BLCA, BRCA, CESC, 
CHOL, COAD, COADREAD, DLBC, ESCA, GBMLGG, HNSC, KICH, KIPAN, KIRC, 
KIRP, LGG, LIHC, LUAD, LUSC, OV, PAAD, PCPG, PRAD, READ, SARC, SKCM,
STAD, STES, TGCT, THCA, THYM, UCEC, UCS, UVM}

\item{exp1}{Custom expression data matrix. Columns must be genes and rows
must be patients. If gene names are provided as header, no need to 
redefine the headers(labels) of the expression data.}

\item{exp2}{Custom expression data matrix. Columns must be genes and rows
must be patients. If gene names are provided as header, no need to 
redefine the headers(labels) of the expression data.}

\item{label1}{Gene names of the custom exp1 expression data. If it is not 
provided, column name of the exp1 data will be taken.}

\item{label2}{Gene names of the custom exp2 expression data. If it is not 
provided, column name of the exp2 data will be taken.}

\item{corrMethod}{Correlation coeffient method that will be used for 
evaluation. Possible values are "pearson", "kendall", "spearman"}

\item{varCutoff}{Variance cutt off that genes have less variance than this
value will be trimmed}

\item{minAbsCor}{Cut-off value for the Pearson correlation coefficient of 
the miRNA-mRNA}

\item{pcut}{P-value cut off for the correlation values}

\item{alternate}{Holds the alternative hypothesis and "two.sided", "greater"
or "less" are the possible values.}

\item{isUnionCorGene}{Boolean value that shows whether union of the output 
of the co-expression analysis and the other analysis should be 
considered}

\item{conf}{Confidence level for the returned confidence interval. It is 
only used for the Pearson correlation coefficient if there are at least
4 complete pairs of observations.}

\item{databaseFile}{Path of miRcancer.db file}

\item{isGeneEnrich}{Boolean value whether gene enrichment should be 
performed}
}
\value{
Pathway enrichment object of the given input
}
\description{
Given gene regions that fall in the given upstream and downstream region of
mRNAs of interest, pathway enrichment analysis is carried out
}
\examples{

regionNC <-  readbed(dm_file = ncRegion,isText = FALSE)
ncPath<-geneRegionPathwayEnricher(region = regionNC, 
                                  org_assembly = 'hg38', 
                                  near = TRUE, 
                                  upstream =1000,
                                  downstream = 0, 
                                  pathwayType = 'kegg')
}
