% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionBased.R
\name{getmiRNACount}
\alias{getmiRNACount}
\title{Get TCGA miRNAseq expression of miRNA genes for the given cancer}
\usage{
getmiRNACount(mirnagene, cancer, databaseFile)
}
\arguments{
\item{mirnagene}{Data frame of the mature format}

\item{cancer}{Name of the TCGA project code such as 'BRCA' that is
analyzed for miRNA-mRNA correlation}

\item{databaseFile}{Path of miRcancer.db file}
}
\value{
Data frame of the raw read count of the given miRNA genes
      for different patients
}
\description{
Get TCGA miRNAseq expression of miRNA genes for the given cancer
}
\examples{

database<- "filePath//miRCancer.db"
\dontrun{
brcaCorr<- getmiRNACount(mirnagene = brain_mirna,
cancer = 'BRCA', databaseFile = database)
}

}
