% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNearest.R
\name{getNearToIntron}
\alias{getNearToIntron}
\title{Get only those neighbouring genes that fall within intron region}
\usage{
getNearToIntron(bedfile, upstream, downstream, hg)
}
\arguments{
\item{bedfile}{Bed file}

\item{upstream}{upstream distance}

\item{downstream}{downstream distance}

\item{hg}{genomee assembly of interest for the analysis. Possible assemblies are "mm10" for mouse, "dre10" for zebrafish, "rn6" for rat, "dm6" for fruit fly, "ce11" for worm, "sc3" for yeast, "hg19" and "hg38" for human}
}
\value{
genes
}
\description{
Get only those neighbouring genes that fall within intron region
}
\examples{

regionNC <-  readbed(dm_file = ncRegion,isText = FALSE)

\dontrun{

r<-getNearToExon(bedfile = regionNC,upstream = 1000, downstream = 2000,hg = 'hg19')
}

}
