% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getReactomeDiagram}
\alias{getReactomeDiagram}
\title{Display the enriched Reactome diagram of the given Reactome pathway id. This function is specific to only one pathway id and identifies the enriched genes in the diagram.}
\usage{
getReactomeDiagram(mrnaObject, pathway, imageFormat)
}
\arguments{
\item{mrnaObject}{Output of enrichment results}

\item{pathway}{Reactome pathway term}

\item{imageFormat}{Image format of the diagram. Possible image formats are 'png', 'svg'}
}
\value{
Shows reactome diagram marked with an enriched genes in a browser
}
\description{
Display the enriched Reactome diagram of the given Reactome pathway id. This function is specific to only one pathway id and identifies the enriched genes in the diagram.
}
\examples{

ncRNAPathway<-mirnaPathwayEnricher(gene = brain_mirna, hg='hg19',
               target = TRUE, min=5, pathwayType = 'reactome',  pAdjust = "none")
getReactomeDiagram(mrnaObject = ncRNAPathway, pathway = ncRNAPathway@ID[1],
                   imageFormat = 'svg')



}
