% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNearest.R
\name{getUCSC}
\alias{getUCSC}
\title{Get nearest genes for the window of the upstream/downstream region.}
\usage{
getUCSC(bedfile, upstream, downstream, hg)
}
\arguments{
\item{bedfile}{Bed formated input gene regions}

\item{upstream}{Maximum upstream distance from the transcription start region of the input gene}

\item{downstream}{Maximum downstream distance from the transcription end region of the input gene}

\item{hg}{genomee assembly of interest for the analysis. Possible assemblies are "mm10" for mouse, "dre10" for zebrafish, "rn6" for rat, "dm6" for fruit fly, "ce11" for worm, "sc3" for yeast, "hg19" and "hg38" for human}
}
\value{
genes
}
\description{
When downstream = 0 / upstream = 0, function converts bed formated regions to HUGO genes
}
\examples{

regionNC <-  readbed(dm_file = ncRegion,isText = FALSE)
 \dontrun{

neighbour <- getUCSC(bedfile = regionNC, upstream = 1000, downstream = 1000,hg = 'hg19')

}
}
