% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{goslim_yeast_p}
\alias{goslim_yeast_p}
\title{Broad, high level GO terms from the biological process ontology selected and maintained by the Saccharomyces Genome Database}
\format{Not Available}
\source{
\url{http://geneontology.org/docs/go-subset-guide/}
}
\usage{
goslim_yeast_p
}
\description{
Broad, high level GO terms from the biological process ontology selected and maintained by the Saccharomyces Genome Database
}
\examples{
data(goslim_yeast_p)
}
\keyword{datasets}
