% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{h_c}
\alias{h_c}
\title{GO-term annotation database from the cellular component ontology for the human genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
h_c
}
\description{
GO-term annotation database from the cellular component ontology for the human genes
}
\examples{
data(h_c)
}
\keyword{datasets}
