% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{h_f}
\alias{h_f}
\title{GO-term annotation database from the molecular function ontology for the human genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
h_f
}
\description{
GO-term annotation database from the molecular function ontology for the human genes
}
\examples{
data(h_f)
}
\keyword{datasets}
