% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{h_p}
\alias{h_p}
\title{GO-term annotation database from the biological process ontology for the human genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
h_p
}
\description{
GO-term annotation database from the biological process ontology for the human genes
}
\examples{
data(h_p)
}
\keyword{datasets}
