% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNearest.R
\name{listTAD}
\alias{listTAD}
\title{List cell line of the given topological domain regions}
\usage{
listTAD(TADName)
}
\arguments{
\item{TADName}{input TAD regions}
}
\value{
cell line of the input tad data
}
\description{
List cell line of the given topological domain regions
}
\examples{
\dontrun{
listTAD(TADName = tad_hg19)
}

}
