% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{mirna}
\alias{mirna}
\title{Brain miRNA expression retrieved from the TCGA}
\format{
Not Available
}
\source{
\url{https://www.gencodegenes.org/}
}
\usage{
mirna
}
\description{
Brain miRNA expression retrieved from the TCGA
}
\examples{
data(mirna)
}
\keyword{datasets}
