% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirnaEnrichment.R
\name{mirnaRegionPathwayEnricher}
\alias{mirnaRegionPathwayEnricher}
\title{Pathway enrichments of the microRNA regions with mRNAs that fall in the given upstream/downstream regions of the microRNA genes}
\usage{
mirnaRegionPathwayEnricher(region, hg, upstream = 10000,
  downstream = 10000, searchRegion = "all", pCut = 0.05,
  pAdjCut = 0.05, pAdjust = "BH", min = 5, pathwayType = "kegg",
  near = FALSE, target = FALSE, isTADSearch = FALSE,
  TAD = tad_hg19, cellline = "all", gmtName = "", isSymbol = FALSE,
  express = FALSE, isCustomExp = FALSE, cancer, exp1, exp2,
  label1 = "", label2 = "", corrMethod = "pearson",
  varCutoff = 0.0025, minAbsCor = 0.3, pcut = 0.05,
  alternate = "greater", isUnionCorGene = FALSE, conf = 0.95,
  databaseFile, isGeneEnrich = FALSE)
}
\arguments{
\item{region}{MiRNA region in a bed format}

\item{hg}{Genome assembly of interest for the analysis. Possible assemblies are "mm10" for mouse, "dre10" for zebrafish, "rn6" for rat, "dm6" for fruit fly, "ce11" for worm, "sc3" for yeast, "hg19" and "hg38" for human}

\item{upstream}{Upstream distance from the transcription start position}

\item{downstream}{Downstream distance from the transcription end position}

\item{searchRegion}{Search space of the cis-region. Possible values are "all", "exon", "intron"}

\item{pCut}{Threshold value for the pvalue. Default value is 0.05}

\item{pAdjCut}{Cutoff value for the adjusted p-values using one of given method. Default value is 0.05.}

\item{pAdjust}{Methods of the adjusted p-values. Possible methods are "bonferroni", "holm", "BH"(default)}

\item{min}{Minimum number of genes that are required for enrichment. By default, it is set to 5.}

\item{pathwayType}{Pathway database for enrichment. Possible values are 'reactome' for Reactome, 'kegg' for KEGG, 'wiki' for WikiPathways, 'other' for custom database}

\item{near}{Boolean value presents whether cis-neighbourhood should be considered in the analysis}

\item{target}{Boolean value shows whether miRNA target prediction should be performed}

\item{isTADSearch}{Boolean value that shows whether TAD analysis is performed. This value has to be TRUE for TAD analysis.}

\item{TAD}{TAD genomic regions for the species. Predefined TAD regions or any new TAD regions can be used for the analysis. TAD regions must be formated as GRanges object. Predefined TAD regions are 'tad_hg19', 'tad_hg38', 'tad_mm10', 'tad_dmel' for hg19, hg38, mm9 and dm6 assembly, respectively.}

\item{cellline}{Cell lines for TAD regions}

\item{gmtName}{Custom pathway gmt file}

\item{isSymbol}{Boolean variable that hold the gene format of the gmt file. If it is set as TRUE, gene format of the gmt file should be symbol. Otherwise, gene format should be ENTREZ ID. By default, it is FALSE.}

\item{express}{Boolean variable whether co-expression analysis is performed. If this option is set to TRUE, co-expression analysis will be performed.}

\item{isCustomExp}{Boolean variable whether co-expression analysis with custom data will be performed. When this option is set, exp1 and exp2 parameters must be defined.}

\item{cancer}{Defines the name of the TCGA project code such as 'BRCA' for correlation analysis. Possible cancer types ACC, BLCA, BRCA, CESC, CHOL, COAD, COADREAD, DLBC, ESCA, GBMLGG, HNSC, KICH, KIPAN, KIRC, KIRP, LGG, LIHC, LUAD, LUSC, OV, PAAD, PCPG, PRAD, READ, SARC, SKCM, STAD, STES, TGCT, THCA, THYM, UCEC, UCS, UVM}

\item{exp1}{Custom expression data matrix. Columns must be genes and rows must be patients. If gene names are provided as header, no need to redefine the headers(labels) of the expression data.}

\item{exp2}{Custom expression data matrix. Columns must be genes and rows must be patients. If gene names are provided as header, no need to redefine the headers(labels) of the expression data.}

\item{label1}{Gene names of the custom exp1 expression data. If it is not provided, column name of the exp1 data will be taken.}

\item{label2}{Gene names of the custom exp2 expression data. If it is not provided, column name of the exp2 data will be taken.}

\item{corrMethod}{Correlation coeffient method that will be used for evaluation. Possible values are "pearson", "kendall", "spearman"}

\item{varCutoff}{Variance cutt off that genes have less variance than this value will be trimmed}

\item{minAbsCor}{Cut-off value for the Pearson correlation coefficient of the miRNA-mRNA}

\item{pcut}{P-value cut off for the correlation values}

\item{alternate}{Holds the alternative hypothesis and "two.sided", "greater" or "less" are the possible values.}

\item{isUnionCorGene}{Boolean value that shows whether union of the output of the co-expression analysis and the other analysis should be considered}

\item{conf}{Confidence level for the returned confidence interval. It is only used for the Pearson correlation coefficient if there are at least 4 complete pairs of observations.}

\item{databaseFile}{Path of miRcancer.db file}

\item{isGeneEnrich}{Boolean value whether gene enrichment should be performed}
}
\value{
miRNA pathway enrichment object for the given input
}
\description{
Pathway enrichments of the microRNA regions with mRNAs that fall in the given upstream/downstream regions of the microRNA genes
}
\examples{

regionNC <-  readbed(dm_file = ncRegion,isText = FALSE)
\dontrun{
a<- mirnaRegionGOEnricher(region = regionNC, hg = 'hg19', near = TRUE)
}

}
