% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{mouse_c}
\alias{mouse_c}
\title{GO-term annotation database from the cellular component ontology for the mouse genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
mouse_c
}
\description{
GO-term annotation database from the cellular component ontology for the mouse genes
}
\examples{
data(mouse_c)
}
\keyword{datasets}
