% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{mouse_f}
\alias{mouse_f}
\title{GO-term annotation database from the molecular function ontology for the mouse genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
mouse_f
}
\description{
GO-term annotation database from the molecular function ontology for the mouse genes
}
\examples{
data(mouse_f)
}
\keyword{datasets}
