% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{mouse_p}
\alias{mouse_p}
\title{GO-term annotation database from the biological process ontology for the mouse genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
mouse_p
}
\description{
GO-term annotation database from the biological process ontology for the mouse genes
}
\examples{
data(mouse_p)
}
\keyword{datasets}
