% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{mrna}
\alias{mrna}
\title{Brain mRNA expression retrieved from the TCGA}
\format{Not Available}
\source{
\url{https://www.gencodegenes.org/}
}
\usage{
mrna
}
\description{
Brain mRNA expression retrieved from the TCGA
}
\examples{
data(mrna)
}
\keyword{datasets}
