% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{ncRegion}
\alias{ncRegion}
\title{Differentially expressed non-coding gene regions}
\format{
Not Available
}
\source{
\url{http://resource.psychencode.org/}
}
\usage{
ncRegion
}
\description{
Differentially expressed non-coding gene regions
}
\examples{
data(ncRegion)
}
\keyword{datasets}
