% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\docType{class}
\name{NoRCE-class}
\alias{NoRCE-class}
\title{An S4 class to represent enrichment}
\description{
An S4 class to represent enrichment
}
\section{Slots}{

\describe{
\item{\code{ID}}{factor}

\item{\code{Term}}{factor}

\item{\code{geneList}}{factor}

\item{\code{ncGeneList}}{factor}

\item{\code{pvalue}}{factor}

\item{\code{pAdj}}{factor}

\item{\code{GeneRatio}}{factor}

\item{\code{BckRatio}}{factor}
}}

