% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNearest.R
\name{packageCheck}
\alias{packageCheck}
\title{Check the package availability for the given assembly}
\usage{
packageCheck(pkg)
}
\arguments{
\item{pkg}{Required packages}
}
\value{
return install packages
}
\description{
Check the package availability for the given assembly
}
