% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{pathwayEnrichment}
\alias{pathwayEnrichment}
\title{For a given gmt file of a specific pathway database, pathway enrichment can be performed. Function supports Entrez ID and symbol based gmt file.}
\usage{
pathwayEnrichment(genes, gmtFile, hg, pCut = 0.05, pAdjCut = 0.05,
  pAdjust = "BH", isSymbol, min = 5, isGeneEnrich = FALSE)
}
\arguments{
\item{genes}{Input genes}

\item{gmtFile}{File path of the gmt file}

\item{hg}{Genome assembly of interest for the analysis. Possible assemblies are "mm10" for mouse, "dre10" for zebrafish, "rn6" for rat, "dm6" for fruit fly, "ce11" for worm, "sc3" for yeast, "hg19" and "hg38" for human}

\item{pCut}{Threshold value for the pvalue. Default value is 0.05}

\item{pAdjCut}{Cutoff value for the adjusted p-values using one of given method. Default value is 0.05.}

\item{pAdjust}{Methods of the adjusted p-values. Possible methods are "bonferroni", "holm", "BH"(default)}

\item{isSymbol}{Boolean value that controls the gene formats. If it is TRUE, gene format of the gmt file should be symbol. Otherwise, gene format must be ENTREZ ID.}

\item{min}{Minimum number of genes that are required for enrichment. By default, it is set to 5.}

\item{isGeneEnrich}{Boolean value whether gene enrichment should be performed}
}
\value{
Pathway Enrichment
}
\description{
For a given gmt file of a specific pathway database, pathway enrichment can be performed. Function supports Entrez ID and symbol based gmt file.
}
\examples{
tmpGene <- breastmRNA[1:100,]
\dontrun{
br_enr<-pathwayEnrichment(genes = tmpGene,hg='hg19',
    gmtFile = 'Human_AllPathways_February_01_2019_symbol.gmt',isSymbol = TRUE)
}
}
