% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{rat_f}
\alias{rat_f}
\title{GO-term annotation database from the molecular function ontology for the rat genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
rat_f
}
\description{
GO-term annotation database from the molecular function ontology for the rat genes
}
\examples{
data(rat_f)
}
\keyword{datasets}
