% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{rat_p}
\alias{rat_p}
\title{GO-term annotation database from the biological process ontology for the rat genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
rat_p
}
\description{
GO-term annotation database from the biological process ontology for the rat genes
}
\examples{
data(rat_p)
}
\keyword{datasets}
