% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNearest.R
\name{readbed}
\alias{readbed}
\title{Convert txt file or data frame to bed format. First three columns should be chr number, start and end, respectively.}
\usage{
readbed(dm_file, isText = TRUE)
}
\arguments{
\item{dm_file}{Bed formated txt file or data frame}

\item{isText}{Boolean value that holds whether input data is txt file. If it is TRUE, input file has to be txt.}
}
\value{
Bed file
}
\description{
Convert txt file or data frame to bed format. First three columns should be chr number, start and end, respectively.
}
\examples{

#txt formatted data
data("ncRegion")
\dontrun{

#Write data to a txt file
write.table(ncRegion,paste("ncRegion.txt"),sep = '\\t',row.names = FALSE,col.names = FALSE)

#Convert txt file that just created to bed formatted data
regionNC <-readbed("ncRegion.txt")

#Directly convert data frame to bed format
regionNC <- readbed(dm_file = ncRegion,isText = FALSE)

}
}
