% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{readGMT}
\alias{readGMT}
\title{Import gmt formatted pathway file to the NoRCE}
\usage{
readGMT(gmtName, isSymbol = FALSE)
}
\arguments{
\item{gmtName}{Custom pathway gmt file}

\item{isSymbol}{Boolean variable that hold the gene format of the gmt file.
If it is set as TRUE, gene format of the gmt file should be symbol.
Otherwise, gene format should be ENTREZ ID. By default, it is FALSE.}
}
\value{
return gmt file
}
\description{
Import gmt formatted pathway file to the NoRCE
}
