% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{sc3}
\alias{sc3}
\title{UCSC data information for the yeast}
\format{Not Available}
\source{
\url{https://genome.ucsc.edu/}
}
\usage{
sc3
}
\description{
UCSC data information for the yeast
}
\examples{
data(sc3)
}
\keyword{datasets}
