% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{slimb}
\alias{slimb}
\title{broad, high level GO terms from the Biological Process ontology selected and maintained by the Gene Ontology Consortium}
\format{Not Available}
\source{
\url{http://geneontology.org/docs/go-subset-guide/}
}
\usage{
slimb
}
\description{
broad, high level GO terms from the Biological Process ontology selected and maintained by the Gene Ontology Consortium
}
\examples{
data(slimb)
}
\keyword{datasets}
