% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{slimc}
\alias{slimc}
\title{broad, high level GO terms from the Cellular Component ontology selected and maintained by the Gene Ontology Consortium}
\format{Not Available}
\source{
\url{http://geneontology.org/docs/go-subset-guide/}
}
\usage{
slimc
}
\description{
broad, high level GO terms from the Cellular Component ontology selected and maintained by the Gene Ontology Consortium
}
\examples{
data(slimc)
}
\keyword{datasets}
