% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{slimf}
\alias{slimf}
\title{broad, high level GO terms from the Molecular Function ontology selected and maintained by the Gene Ontology Consortium}
\format{Not Available}
\source{
\url{http://geneontology.org/docs/go-subset-guide/}
}
\usage{
slimf
}
\description{
broad, high level GO terms from the Molecular Function ontology selected and maintained by the Gene Ontology Consortium
}
\examples{
data(slimf)
}
\keyword{datasets}
