% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{tad_dmel}
\alias{tad_dmel}
\title{TAD regions for the fly}
\format{Not Available}
\source{
\url{http://chorogenome.ie-freiburg.
mpg.de/data_sources.html#hi-c_datasets }
}
\usage{
tad_dmel
}
\description{
TAD regions for the fly
}
\examples{
data(tad_dmel)
}
\keyword{datasets}
