% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{tad_hg19}
\alias{tad_hg19}
\title{TAD regions for human hg19 assembly}
\format{Not Available}
\source{
\url{http://promoter.bx.psu.edu/hi-c/publications.html}
}
\usage{
tad_hg19
}
\description{
TAD regions for human hg19 assembly
}
\examples{
data(tad_hg19)
}
\keyword{datasets}
