% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{tad_hg38}
\alias{tad_hg38}
\title{TAD regions for human hg38 assembly}
\format{Not Available}
\source{
\url{http://promoter.bx.psu.edu/hi-c/publications.html}
}
\usage{
tad_hg38
}
\description{
TAD regions for human hg38 assembly
}
\examples{
data(tad_hg38)
}
\keyword{datasets}
