% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{tad_mm10}
\alias{tad_mm10}
\title{TAD regions for mouse}
\format{
Not Available
}
\source{
\url{http://promoter.bx.psu.edu/hi-c/publications.html}
}
\usage{
tad_mm10
}
\description{
TAD regions for mouse
}
\examples{
data(tad_mm10)
}
\keyword{datasets}
