% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{targets_fly}
\alias{targets_fly}
\title{TargetScan database for the fly}
\format{Not Available}
\source{
\url{http://www.targetscan.org}
}
\usage{
targets_fly
}
\description{
TargetScan database for the fly
}
\examples{
data(targets_fly)
}
\keyword{datasets}
