% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{targets_human}
\alias{targets_human}
\title{TargetScan database for the human}
\format{Not Available}
\source{
\url{http://www.targetscan.org}
}
\usage{
targets_human
}
\description{
TargetScan database for the human
}
\examples{
data(targets_human)
}
\keyword{datasets}
