% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{targets_mouse}
\alias{targets_mouse}
\title{TargetScan database for the mouse}
\format{Not Available}
\source{
\url{http://www.targetscan.org}
}
\usage{
targets_mouse
}
\description{
TargetScan database for the mouse
}
\examples{
data(targets_mouse)
}
\keyword{datasets}
