% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{targets_rat}
\alias{targets_rat}
\title{mirmap database for rat}
\format{Not Available}
\source{
\url{https://mirmap.ezlab.org/}
}
\usage{
targets_rat
}
\description{
mirmap database for rat
}
\examples{
data(targets_rat)
}
\keyword{datasets}
