% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{targets_worm}
\alias{targets_worm}
\title{TargetScan database for worm}
\format{Not Available}
\source{
\url{http://www.targetscan.org}
}
\usage{
targets_worm
}
\description{
TargetScan database for worm
}
\examples{
data(targets_worm)
}
\keyword{datasets}
