% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{targets_zebra}
\alias{targets_zebra}
\title{TargetScan database for zebra fish}
\format{Not Available}
\source{
\url{http://www.targetscan.org}
}
\usage{
targets_zebra
}
\description{
TargetScan database for zebra fish
}
\examples{
data(targets_zebra)
}
\keyword{datasets}
