% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{ucsc_ce11}
\alias{ucsc_ce11}
\title{UCSC information for worm}
\format{Not Available}
\source{
\url{https://genome.ucsc.edu/}
}
\usage{
ucsc_ce11
}
\description{
UCSC information for worm
}
\examples{
data(ucsc_ce11)
}
\keyword{datasets}
