% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{ucsc_dm6}
\alias{ucsc_dm6}
\title{UCSC information for fly}
\format{Not Available}
\source{
\url{https://genome.ucsc.edu/}
}
\usage{
ucsc_dm6
}
\description{
UCSC information for fly
}
\examples{
data(ucsc_dm6)
}
\keyword{datasets}
