% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{ucsc_dre10}
\alias{ucsc_dre10}
\title{UCSC information for zebra fish}
\format{Not Available}
\source{
\url{https://genome.ucsc.edu/}
}
\usage{
ucsc_dre10
}
\description{
UCSC information for zebra fish
}
\examples{
data(ucsc_dre10)
}
\keyword{datasets}
