% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{ucsc_hg19}
\alias{ucsc_hg19}
\title{UCSC information for human hg19 assembly}
\format{Not Available}
\source{
\url{https://genome.ucsc.edu/}
}
\usage{
ucsc_hg19
}
\description{
UCSC information for human hg19 assembly
}
\examples{
data(ucsc_hg19)
}
\keyword{datasets}
