% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{ucsc_hg38}
\alias{ucsc_hg38}
\title{UCSC information for human hg38 assembly}
\format{Not Available}
\source{
\url{https://genome.ucsc.edu/}
}
\usage{
ucsc_hg38
}
\description{
UCSC information for human hg38 assembly
}
\examples{
data(ucsc_hg38)
}
\keyword{datasets}
