% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{ucsc_mm10}
\alias{ucsc_mm10}
\title{UCSC information for mouse}
\format{Not Available}
\source{
\url{https://genome.ucsc.edu/}
}
\usage{
ucsc_mm10
}
\description{
UCSC information for mouse
}
\examples{
data(ucsc_mm10)
}
\keyword{datasets}
