% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{ucsc_rn6}
\alias{ucsc_rn6}
\title{UCSC information for rat}
\format{Not Available}
\source{
\url{https://genome.ucsc.edu/}
}
\usage{
ucsc_rn6
}
\description{
UCSC information for rat
}
\examples{
data(ucsc_rn6)
}
\keyword{datasets}
