% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{WikiEnrichment}
\alias{WikiEnrichment}
\title{WikiPathways Enrichment}
\usage{
WikiEnrichment(genes, hg, pCut = 0.05, pAdjCut = 0.05,
  pAdjust = "BH", min = 5)
}
\arguments{
\item{genes}{Input genes}

\item{hg}{Genome assembly of interest for the analysis. Possible assemblies are "mm10" for mouse, "dre10" for zebrafish, "rn6" for rat, "dm6" for fruit fly, "ce11" for worm, "sc3" for yeast, "hg19" and "hg38" for human}

\item{pCut}{Threshold value for the pvalue. Default value is 0.05}

\item{pAdjCut}{Cutoff value for the adjusted p-values using one of given method. Default value is 0.05.}

\item{pAdjust}{Methods of the adjusted p-values. Possible methods are "bonferroni", "holm", "BH"(default)}

\item{min}{Minimum number of genes that are required for enrichment. By default, it is set to 5.}
}
\value{
Wiki Pathway Enrichment
}
\description{
WikiPathways Enrichment
}
\examples{

br_enr<-WikiEnrichment(genes = breastmRNA,hg='hg19')

}
