% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{worm_c}
\alias{worm_c}
\title{GO-term annotation database from the cellular component ontology for the worm genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
worm_c
}
\description{
GO-term annotation database from the cellular component ontology for the worm genes
}
\examples{
data(worm_c)
}
\keyword{datasets}
