% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{worm_f}
\alias{worm_f}
\title{GO-term annotation database from the molecular function ontology for the worm genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
worm_f
}
\description{
GO-term annotation database from the molecular function ontology for the worm genes
}
\examples{
data(worm_f)
}
\keyword{datasets}
