% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{worm_p}
\alias{worm_p}
\title{GO-term annotation database from the biological process ontology for the worm genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
worm_p
}
\description{
GO-term annotation database from the biological process ontology for the worm genes
}
\examples{
data(worm_p)
}
\keyword{datasets}
