% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{yeast_c}
\alias{yeast_c}
\title{GO-term annotation database from the cellular component ontology for the yeast genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
yeast_c
}
\description{
GO-term annotation database from the cellular component ontology for the yeast genes
}
\examples{
data(yeast_c)
}
\keyword{datasets}
