% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{yeast_f}
\alias{yeast_f}
\title{GO-term annotation database from the molecular functional ontology for the yeast genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
yeast_f
}
\description{
GO-term annotation database from the molecular functional ontology for the yeast genes
}
\examples{
data(yeast_f)
}
\keyword{datasets}
