% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{yeast_p}
\alias{yeast_p}
\title{GO-term annotation database from the biological process ontology for the yeast genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
yeast_p
}
\description{
GO-term annotation database from the biological process ontology for the yeast genes
}
\examples{
data(yeast_p)
}
\keyword{datasets}
