% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{zebra_c}
\alias{zebra_c}
\title{GO-term annotation database from the cellular component ontology for the zebra genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
zebra_c
}
\description{
GO-term annotation database from the cellular component ontology for the zebra genes
}
\examples{
data(zebra_c)
}
\keyword{datasets}
