% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{zebra_f}
\alias{zebra_f}
\title{GO-term annotation database from the molecular function ontology for the zebra genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
zebra_f
}
\description{
GO-term annotation database from the molecular function ontology for the zebra genes
}
\examples{
data(zebra_f)
}
\keyword{datasets}
