% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expData.R
\docType{data}
\name{zebra_p}
\alias{zebra_p}
\title{GO-term annotation database from the biological process ontology for the zet genes}
\format{Not Available}
\source{
\url{https://www.ebi.ac.uk/GOA/downloads}
}
\usage{
zebra_p
}
\description{
GO-term annotation database from the biological process ontology for the zet genes
}
\examples{
data(zebra_p)
}
\keyword{datasets}
